Mls 5 (#1)
* Add "spray" request parameter name to getSpray method

* MLS-5 Create a page to use with the URL Shortener

Refactor view controllers to use ViewNameEnum and add URL shortener page

- Update ViewNameEnum with HOME and SHORTENER values
- Add CommonConfs object to store common configuration like site title
- Refactor ViewSprayController to use ViewNameEnum.SPRAY
- Add ViewRootController for root "/" mapping using ViewNameEnum.HOME
- Add ViewShortenerController for "/shortener" page
- Update ThymeleafViewControllerInterface to provide common getModelAndView method
- Refactor Thymeleaf templates to use view name enum
- Add shortener.html template for URL shortener page
- Add JavaScript functions in mt-script.js to copy shortened URL to clipboard
- Update Maven run configuration to enable test profile by default

This commit refactors the view controllers to use the ViewNameEnum for view names instead of hardcoded strings. It adds a URL shortener page at "/shortener". The Thymeleaf templates are updated to work with the refactored controllers and use the view name enums. A new JavaScript file is added with common functions. The Maven run configuration is updated to enable the test profile by default.